+++ CHANGELOG.md
@@ -3,6 +3,10 @@
  * `format_coverage` no longer emits SDK coverage unless --sdk-root is set
    explicitly.
 
+ * Removed support for collecting coverage from old (<1.9.0) Dart SDKs.
+ 
+ * Removed deprecated `Resolver.pkgRoot`.
+
 ###0.6.5
 
  * Fixed early collection bug when --wait-paused is set.